[CL 4161603 by Jack Porter in Dev-Mobile branch]

+++ Engine/Source/Runtime/D3D12RHI/Private/D3D12DescriptorCache.h
@@ -117,7 +117,7 @@ struct FD3D12UniqueDescriptorTable
 
 	FORCEINLINE uint32 GetTypeHash(const FD3D12UniqueDescriptorTable& Table)
 	{
-		return uint32(FD3D12PipelineStateCache::HashData((void*)Table.Key.SamplerID, Table.Key.Count * sizeof(Table.Key.SamplerID[0])));
+		return FD3D12PipelineStateCache::HashData((void*)Table.Key.SamplerID, Table.Key.Count * sizeof(Table.Key.SamplerID[0]));
 	}
 
 	FD3D12SamplerArrayDesc Key;
